Move addTodo function to module
