+++ src/index.js
@@ -1,36 +1,18 @@
+/** @format */
+
 //import './style.css';
 import todoStatusUpdate from './todoStatusUpdate.js';
 import dragDrop from './dragDrop.js';
 import deleteCompletedTodoItem from './deleteAll.js';
+import addTodo from './addTodo.js'
 
 const tasks = localStorage.getItem('items')
   ? JSON.parse(localStorage.getItem('items'))
-  : [
-    { index: 0, description: 'Learn webpack', completed: false },
-    { index: 1, description: 'Read 20 pages of book', completed: false },
-    { index: 2, description: 'Create portfolio', completed: false },
-  ];
+  : [];
 
 localStorage.setItem('items', JSON.stringify(tasks));
 const data = JSON.parse(localStorage.getItem('items'));
 
-const addTodoTextBox = () => {
-  const inputText = document.getElementById('input-text');
-  inputText.addEventListener('keyup', (e) => {
-    if (e.keyCode === 13) {
-      const newItem = {
-        index: parseInt(`${tasks.length}`),
-        description: `${inputText.value}`,
-        completed: false,
-      };
-      createList(newItem);
-      tasks.push(newItem);
-      localStorage.setItem('items', JSON.stringify(tasks));
-    }
-  });
-};
-
-addTodoTextBox();
 
 const createList = (todoItem) => {
   if (!todoItem) {
@@ -64,7 +46,7 @@ const createList = (todoItem) => {
   todoStatusUpdate(tasks);
 };
 
-
+addTodo(tasks);
 deleteCompletedTodoItem(tasks);
 data.forEach((object) => {
   createList(object);
